requirements_api.txt
aniso8601==9.0.1
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
flask-mongoengine==1.0.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.23.0
pymongo==3.11.3
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
